refactor: add subfolders in the src/mappers/ folder
to distinguish between api-to-domain and record-to-domain mapping.
