+++ src/mappers/api/AbilityBonusApiToDomainMapper.ts
@@ -1,8 +1,8 @@
-import { IMapper } from '../interfaces/IMapper.js';
-import { AbilityBonusApiDto } from '../types/api/helpers/AbilityBonusApiDto.js';
-import { BaseResourceApiDto } from '../types/api/wrappers/BaseResourceApiDto.js';
-import { AbilityBonus } from '../types/domain/helpers/AbilityBonus.js';
-import { BaseResource } from '../types/domain/wrappers/BaseResource.js';
+import { IMapper } from '../../interfaces/IMapper.js';
+import { AbilityBonusApiDto } from '../../types/api/helpers/AbilityBonusApiDto.js';
+import { BaseResourceApiDto } from '../../types/api/wrappers/BaseResourceApiDto.js';
+import { AbilityBonus } from '../../types/domain/helpers/AbilityBonus.js';
+import { BaseResource } from '../../types/domain/wrappers/BaseResource.js';
 
 export class AbilityBonusApiToDomainMapper
 	implements IMapper<AbilityBonusApiDto, AbilityBonus>
