+++ src/mappers/api/ArmorApiToDomainMapper.ts
@@ -1,11 +1,11 @@
-import { IMapper } from '../interfaces/IMapper.js';
+import { IMapper } from '../../interfaces/IMapper.js';
 import {
 	ArmorApiDto,
 	ArmorClassApiDto,
-} from '../types/api/resources/ArmorApiDto.js';
-import { EquipmentApiDto } from '../types/api/resources/EquipmentApiDto.js';
-import { Armor, ArmorClass } from '../types/domain/resources/Armor.js';
-import { Equipment } from '../types/domain/resources/Equipment.js';
+} from '../../types/api/resources/ArmorApiDto.js';
+import { EquipmentApiDto } from '../../types/api/resources/EquipmentApiDto.js';
+import { Armor, ArmorClass } from '../../types/domain/resources/Armor.js';
+import { Equipment } from '../../types/domain/resources/Equipment.js';
 
 export class ArmorApiToDomainMapper implements IMapper<ArmorApiDto, Armor> {
 	/**
