+++ src/mappers/api/BaseResourceApiToDomainMapper.ts
@@ -1,6 +1,6 @@
-import { IMapper } from '../interfaces/IMapper.js';
-import { BaseResourceApiDto } from '../types/api/wrappers/BaseResourceApiDto.js';
-import { BaseResource } from '../types/domain/wrappers/BaseResource.js';
+import { IMapper } from '../../interfaces/IMapper.js';
+import { BaseResourceApiDto } from '../../types/api/wrappers/BaseResourceApiDto.js';
+import { BaseResource } from '../../types/domain/wrappers/BaseResource.js';
 
 /**
  * Maps base resource API DTOs to domain model base resources.
