+++ src/mappers/api/ProficiencyApiToDomainMapper.ts
@@ -1,8 +1,8 @@
-import { IMapper } from '../interfaces/IMapper.js';
-import { ProficiencyApiDto } from '../types/api/resources/ProficiencyApiDto.js';
-import { BaseResourceApiDto } from '../types/api/wrappers/BaseResourceApiDto.js';
-import { Proficiency } from '../types/domain/resources/Proficiency.js';
-import { BaseResource } from '../types/domain/wrappers/BaseResource.js';
+import { IMapper } from '../../interfaces/IMapper.js';
+import { ProficiencyApiDto } from '../../types/api/resources/ProficiencyApiDto.js';
+import { BaseResourceApiDto } from '../../types/api/wrappers/BaseResourceApiDto.js';
+import { Proficiency } from '../../types/domain/resources/Proficiency.js';
+import { BaseResource } from '../../types/domain/wrappers/BaseResource.js';
 
 export class ProficiencyApiToDomainMapper
 	implements IMapper<ProficiencyApiDto, Proficiency>
