+++ src/mappers/api/SubclassApiToDomainMapper.ts
@@ -1,8 +1,8 @@
-import { IMapper } from '../interfaces/IMapper.js';
-import { SubclassApiDto } from '../types/api/resources/SubclassApiDto.js';
-import { BaseResourceApiDto } from '../types/api/wrappers/BaseResourceApiDto.js';
-import { Subclass } from '../types/domain/resources/Subclass.js';
-import { BaseResource } from '../types/domain/wrappers/BaseResource.js';
+import { IMapper } from '../../interfaces/IMapper.js';
+import { SubclassApiDto } from '../../types/api/resources/SubclassApiDto.js';
+import { BaseResourceApiDto } from '../../types/api/wrappers/BaseResourceApiDto.js';
+import { Subclass } from '../../types/domain/resources/Subclass.js';
+import { BaseResource } from '../../types/domain/wrappers/BaseResource.js';
 
 export class SubclassApiToDomainMapper
 	implements IMapper<SubclassApiDto, Subclass>
