+++ src/wiring/dependencies.ts
@@ -1,28 +1,23 @@
-import { AbilityBonusApiToDomainMapper } from '../mappers/AbilityBonusApiToDomainMapper.js';
-import { AbilityScoreApiToDomainMapper } from '../mappers/AbilityScoreApiToDomainMapper.js';
-import { AlignmentApiToDomainMapper } from '../mappers/AlignmentApiToDomainMapper.js';
-import { ArmorApiToDomainMapper } from '../mappers/ArmorApiToDomainMapper.js';
-import { BackgroundApiToDomainMapper } from '../mappers/BackgroundApiToDomainMapper.js';
-import { BaseResourceApiToDomainMapper } from '../mappers/BaseResourceApiToDomainMapper.js';
-import { ChoiceApiToDomainMapper } from '../mappers/ChoiceApiToDomainMapper.js';
-import { ClassLevelApiToDomainMapper } from '../mappers/ClassLevelApiToDomainMapper.js';
-import { ClassApiToDomainMapper } from '../mappers/ClassApiToDomainMapper.js';
-import { EquipmentCategoryApiToDomainMapper } from '../mappers/EquipmentCategoryApiToDomainMapper.js';
-import { EquipmentApiToDomainMapper } from '../mappers/EquipmentApiToDomainMapper.js';
-import { FeatureApiToDomainMapper } from '../mappers/FeatureApiToDomainMapper.js';
-import { LanguageApiToDomainMapper } from '../mappers/LanguageApiToDomainMapper.js';
-import { ProficiencyApiToDomainMapper } from '../mappers/ProficiencyApiToDomainMapper.js';
-import { RaceApiToDomainMapper } from '../mappers/RaceApiToDomainMapper.js';
-import { SkillApiToDomainMapper } from '../mappers/SkillApiToDomainMapper.js';
-import { SubclassApiToDomainMapper } from '../mappers/SubclassApiToDomainMapper.js';
-import { SubraceApiToDomainMapper } from '../mappers/SubraceApiToDomainMapper.js';
-import { TraitApiToDomainMapper } from '../mappers/TraitApiToDomainMapper.js';
-import { WeaponApiToDomainMapper } from '../mappers/WeaponApiToDomainMapper.js';
-import { BaseResourceRecordToDomainMapper } from '../mappers/BaseResourceRecordToDomainMapper.js';
-import { AbilityBonusRecordToDomainMapper } from '../mappers/AbilityBonusRecordToDomainMapper.js';
-import { ChoiceRecordToDomainMapper } from '../mappers/ChoiceRecordToDomainMapper.js';
-import { TraitRecordToDomainMapper } from '../mappers/TraitRecordToDomainMapper.js';
-import { RaceRecordToDomainMapper } from '../mappers/RaceRecordToDomainMapper.js';
+import { AbilityBonusApiToDomainMapper } from '../mappers/api/AbilityBonusApiToDomainMapper.js';
+import { AbilityScoreApiToDomainMapper } from '../mappers/api/AbilityScoreApiToDomainMapper.js';
+import { AlignmentApiToDomainMapper } from '../mappers/api/AlignmentApiToDomainMapper.js';
+import { ArmorApiToDomainMapper } from '../mappers/api/ArmorApiToDomainMapper.js';
+import { BackgroundApiToDomainMapper } from '../mappers/api/BackgroundApiToDomainMapper.js';
+import { BaseResourceApiToDomainMapper } from '../mappers/api/BaseResourceApiToDomainMapper.js';
+import { ChoiceApiToDomainMapper } from '../mappers/api/ChoiceApiToDomainMapper.js';
+import { ClassLevelApiToDomainMapper } from '../mappers/api/ClassLevelApiToDomainMapper.js';
+import { ClassApiToDomainMapper } from '../mappers/api/ClassApiToDomainMapper.js';
+import { EquipmentCategoryApiToDomainMapper } from '../mappers/api/EquipmentCategoryApiToDomainMapper.js';
+import { EquipmentApiToDomainMapper } from '../mappers/api/EquipmentApiToDomainMapper.js';
+import { FeatureApiToDomainMapper } from '../mappers/api/FeatureApiToDomainMapper.js';
+import { LanguageApiToDomainMapper } from '../mappers/api/LanguageApiToDomainMapper.js';
+import { ProficiencyApiToDomainMapper } from '../mappers/api/ProficiencyApiToDomainMapper.js';
+import { RaceApiToDomainMapper } from '../mappers/api/RaceApiToDomainMapper.js';
+import { SkillApiToDomainMapper } from '../mappers/api/SkillApiToDomainMapper.js';
+import { SubclassApiToDomainMapper } from '../mappers/api/SubclassApiToDomainMapper.js';
+import { SubraceApiToDomainMapper } from '../mappers/api/SubraceApiToDomainMapper.js';
+import { TraitApiToDomainMapper } from '../mappers/api/TraitApiToDomainMapper.js';
+import { WeaponApiToDomainMapper } from '../mappers/api/WeaponApiToDomainMapper.js';
 import { BaseResourceRepository } from '../repositories/BaseResourceRepository.js';
 import { ClassLevelRepository } from '../repositories/ClassLevelRepository.js';
 import { FeatureRepository } from '../repositories/FeatureRepository.js';
@@ -31,6 +26,11 @@ import { TraitRepository } from '../repositories/TraitRepository.js';
 import { CacheService } from '../services/CacheService.js';
 import { LocalStorageService } from '../services/LocalStorageService.js';
 import { SrdApiService } from '../services/SrdApiService.js';
+import { AbilityBonusRecordToDomainMapper } from '../mappers/record/AbilityBonusRecordToDomainMapper.js';
+import { BaseResourceRecordToDomainMapper } from '../mappers/record/BaseResourceRecordToDomainMapper.js';
+import { ChoiceRecordToDomainMapper } from '../mappers/record/ChoiceRecordToDomainMapper.js';
+import { RaceRecordToDomainMapper } from '../mappers/record/RaceRecordToDomainMapper.js';
+import { TraitRecordToDomainMapper } from '../mappers/record/TraitRecordToDomainMapper.js';
 
 /**
  * Dependency injection container for the entire application.
